requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1

llama-index-core==0.12.24.post1 # Explicitly set llama-index-core
llama-index-vector-stores-pinecone==0.6.0 # Compatible with core >=0.12.0
llama-index-embeddings-huggingface==0.4.0 # Updated to be compatible with core >=0.12.0
llama-index-llms-google-genai==0.1.5

pinecone-client==3.2.2
huggingface-hub==0.23.2
transformers==4.42.0
numpy<2.0
google-generativeai==0.8.5
